[gardening] Move switch building utility code from GenEnum.cpp to SwitchBuilder.h
This code is utility code that is just thrown around in the middle of the enum
emission code and makes it difficult to follow. It is something that is large
enough really to stand on its on.
